Regression AutoML Learner

The LearnerRegrAuto is an automated machine learning (AutoML) system for regression tasks. It combines preprocessing, a switch between multiple learners, and hyperparameter tuning to find the best model for the given task.

Debugging

Set options(bbotk.debug) to run the tuning in the in the main session. Set encapsulate_learner = FALSE to remove encapsulation of the learner. Set encapsulate_mbo = FALSE to catch no errors in mbo.

Parameters

learner_timeout

(integer(1))
Timeout for training and predicting with a single learner.

n_threads

(integer(1))
Number of threads to use for model training.

memory_limit

(integer(1))
Memory limit for model training in MB.

devices

(character())
Devices to use for model training. Possible values are "cpu" and "cuda". If "cuda", the learner will be trained on a GPU.

large_data_size

(integer(1))
Threshold value for the data set size from which special rules apply. Only the learners specified in large_data_learner_ids will be considered. These learners can use up to large_data_nthread threads.

small_data_size

(integer(1))
Threshold value for the data set size from which special rules apply.

small_data_resampling

(mlr3::Resampling)
Resampling strategy to use for model training on small data sets.

initial_design_default

(logical(1))
Whether to use the default design of the learner.

initial_design_set

(integer(1))
Number of points to use for the initial design set.

initial_design_size

(integer(1))
Size of the random, sobol or lhs initial design.

initial_design_type

(character(1))
Type of the initial design used for mbo. Possible values are "lhs", "sobol", "random". "lhs" uses a Latin Hypercube Sampling design. "sobol" uses a Sobol sequence design. "random" uses a random design.

initial_design_fraction

(numeric(1))
Fraction of the budget to use for the initial design.

resampling

(mlr3::Resampling)
Resampling strategy used for tuning.

terminator

(bbotk::Terminator)
Terminator criterion for tuning.

measure

(mlr3::Measure)
Measure used for tuning.

callbacks

(mlr3tuning::CallbackAsyncTuning)
Callbacks used for tuning.

store_benchmark_result

(logical(1))
Whether to store the benchmark result.

store_models

(logical(1))
Whether to store the models.

encapsulate_learner

(logical(1))
Whether to encapsulate the learner. Change to FALSE to debug.

encapsulate_mbo

(logical(1))
Whether to encapsulate the tuning. Change to FALSE to debug.

check_learners

(logical(1))
Whether to check if the learners are compatible with the task. Change to FALSE to debug.

Super classes

mlr3::Learner -> mlr3automl::LearnerAuto -> LearnerRegrAuto

Methods

Public methods

LearnerRegrAuto$new()

LearnerRegrAuto$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrAuto$new(id = "regr.auto", learner_ids, rush = NULL)

Arguments

id

(character(1))
Identifier for the new instance.

learner_ids

(character())
Learner that should be used.

rush

rush::Rush
Rush instance.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrAuto$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.